requirements.txt
attrs==17.3.0
blessings==1.6
bpython==0.16
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
croniter==0.3.20
curtsies==0.2.11
dj-database-url==0.4.2
Django==2.0
django-choices==1.6.0
django-extensions==1.9.8
djangorestframework==3.7.3
future==0.16.0
greenlet==0.4.12
gunicorn==19.7.1
idna==2.6
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.2
Pygments==2.2.0
pytest==3.3.1
pytest-django==3.1.2
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
typing==3.6.2
urllib3==1.22
wcwidth==0.1.7
whitenoise==3.3.1
